4.21*: bloglist expansion, step9

+++ controllers/blogs.js
@@ -35,8 +35,14 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id) 
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const blogToDelete = await Blog.findById(request.params.id)
+    if (decodedToken.id === blogToDelete.user.toString()) {
+        await Blog.findByIdAndRemove(request.params.id) 
+        return response.status(204).end()
+    }
+
+    response.status(401).json({ error: 'token missing or invalid' })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
